Move getSupportedPrisonIds() to orchestration API client
