+++ integration_tests/mockApis/orchestration.ts
@@ -1,6 +1,7 @@
 import { SuperAgentRequest } from 'superagent'
 import { stubFor } from './wiremock'
 import { PrisonerProfile, VisitHistoryDetails } from '../../server/data/orchestrationApiTypes'
+import TestData from '../../server/routes/testutils/testData'
 
 export default {
   stubVisitHistory: (visitHistoryDetails: VisitHistoryDetails): SuperAgentRequest => {
@@ -34,6 +35,19 @@ export default {
         status: 200,
         headers: { 'Content-Type': 'application/json;charset=UTF-8' },
         jsonBody: profile,
+      },
+    })
+  },
+  stubSupportedPrisonIds: (): SuperAgentRequest => {
+    return stubFor({
+      request: {
+        method: 'GET',
+        url: '/orchestration/config/prisons/supported',
+      },
+      response: {
+        status: 200,
+        headers: { 'Content-Type': 'application/json;charset=UTF-8' },
+        jsonBody: TestData.supportedPrisonIds(),
       },
     })
   },
