+++ server/data/orchestrationApiClient.ts
@@ -9,11 +9,22 @@ export default class OrchestrationApiClient {
     this.restClient = new RestClient('orchestrationApiClient', config.apis.orchestration as ApiConfig, token)
   }
 
+  // orchestration-visits-controller
+
   async getVisitHistory(reference: string): Promise<VisitHistoryDetails> {
     return this.restClient.get({ path: `/visits/${reference}/history` })
   }
 
+  // prisoner-profile-controller
+
   async getPrisonerProfile(prisonId: string, prisonerId: string): Promise<PrisonerProfile> {
     return this.restClient.get({ path: `/prisoner/${prisonId}/${prisonerId}/profile` })
+  }
+
+  // orchestration-prisons-config-controller
+  async getSupportedPrisonIds(): Promise<string[]> {
+    return this.restClient.get({
+      path: '/config/prisons/supported',
+    })
   }
 }
